resolve conflict

+++ src/modules/semester/semester.service.ts
@@ -3,13 +3,24 @@ import { Semester } from "./semester.model";
 
 export class SemesterService {
     private semesterRepository: Repository<Semester>;
-
     constructor(private readonly dataSource: DataSource) {
         this.semesterRepository = this.dataSource.getRepository(Semester);
     }
 
     async createSemester(data) {
         try {
+            const semester = await this.semesterRepository.findOne({
+                where: {
+                    name: data.name
+                }
+            });
+
+            console.log(semester);
+            if(semester) {
+                throw new Error("Semester is existed in system")
+            }
+
+            // check condition date khong duoc trong date da tao
             const newSemester = this.semesterRepository.create({
                 ...data,
                 isActive: true
@@ -24,15 +35,28 @@ export class SemesterService {
         try {
             return await this.semesterRepository.find();
         } catch (error) {
-            return error;
+            throw new Error(error);
         }
     }
 
     async deleteSemester(semesterId: string) {
+        console.log('semesterId', semesterId)
         try {
+            const semester = await this.semesterRepository.findOne({
+                where: {
+                    id: semesterId
+                }
+            });
+
+            console.log(semester);
+            if(!semester) {
+                throw new Error("Semester is not exist")
+            }
+
+
             return await this.semesterRepository.delete(semesterId);
         } catch (error) {
-            return error;
+            throw new Error(error);
         }
     }
 }
